// frontend/src/client/types.gen.ts
export type TaskCreate = {
  title: string
  description: string
  assignee: string
  status?: TaskStatus
  created_at?: string
}

export type TaskPublic = {
  title: string
  description: string
  assignee: string
  status?: TaskStatus
  created_at?: string
  id: string
}

export type TasksPublic = {
  data: Array<TaskPublic>
  count: number
}

export type TaskStatus = "To Do" | "In Progress"